extract extension methods of Filter Descriptors
