+++ src/Common/Utils/ExpressionExtensions.cs
@@ -1,10 +1,9 @@
 using System.Linq.Expressions;
 using System.Reflection;
-using Boilerplate.Common.Data.Querying;
 
 namespace Boilerplate.Common.Utils;
 
-public static class Expressions
+public static class ExpressionExtensions
 {
     private class ParameterExpressionVisitor : ExpressionVisitor
     {
@@ -47,31 +46,6 @@ public static class Expressions
     {
         var negated = Expression.Not(expr.Body);
         return Expression.Lambda<Func<T, bool>>(negated, expr.Parameters);
-    }
-
-    public static Expression<Func<T, bool>> ToExpression<T>(this FilterDescriptor descriptor)
-    {
-        if (descriptor is CompositeFilterDescriptor compositeDescriptor) {
-            return compositeDescriptor.ToExpression<T>();
-        }
-
-        var param = Expression.Parameter(typeof(T));
-        var property = Expression.Property(param, descriptor.Field);
-        var valueParam = Expression.Constant(descriptor.Value);
-        var expression = descriptor.Operator.CreateExpression(property, valueParam);
-
-        return Expression.Lambda<Func<T, bool>>(expression, param);
-    }
-
-    public static Expression<Func<T, bool>> ToExpression<T>(this CompositeFilterDescriptor descriptor)
-    {
-        if (!descriptor.Filters.Any()) {
-            return _ => true;
-        }
-
-        return descriptor.Filters
-            .Select(ToExpression<T>)
-            .Aggregate((prev, next) => descriptor.Logic.CreateExpression(prev, next));
     }
 
     public static Type? GetMemberReturnType(this Expression expression)
